requirements.txt
mcp>=0.2.0
openai>=1.3.0
python-decouple>=3.8
python-dotenv>=1.0.0
httpx>=0.25.0